style: update ChopLogicButton
