+++ src/components/inputs/button/Button.module.scss
@@ -1,0 +1,156 @@
+@use '/src/styles/typography';
+@use '/src/styles/mixins';
+
+.primary {
+  position: relative;
+  border: none;
+  background: transparent;
+  padding: 0;
+  cursor: pointer;
+  outline: none;
+  transition: filter 250ms;
+  width: inherit;
+
+  &__extended {
+    width: 100%;
+  }
+
+  &[disabled] {
+    @include mixins.disabled();
+  }
+
+  &_shadow {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    border-radius: var(--cl-block-border-radius);
+    background: hsl(0deg 0% 0% / 0.25);
+    will-change: transform;
+    transform: translateY(2px);
+    transition: transform 600ms cubic-bezier(0.3, 0.7, 0.4, 1);
+  }
+
+  &_edge {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    border-radius: var(--cl-block-border-radius);
+    background: var(--cl-gradient-shadow);
+  }
+
+  &_front {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: var(--cl-gap-small);
+    position: relative;
+    padding: var(--cl-gap-medium) var(--cl-gap-big);
+    border-radius: var(--cl-block-border-radius);
+    font-size: var(--cl-font-size-base);
+    color: #ffffff;
+    background: var(--cl-font-color-accent);
+    will-change: transform;
+    transform: translateY(-4px);
+    transition: transform 600ms cubic-bezier(0.3, 0.7, 0.4, 1);
+  }
+
+  &:hover {
+    filter: brightness(110%);
+  }
+
+  &:hover &_front {
+    transform: translateY(-6px);
+    transition: transform 250ms cubic-bezier(0.3, 0.7, 0.4, 1.5);
+  }
+
+  &:active &_front {
+    transform: translateY(-2px);
+    transition: transform 34ms;
+  }
+
+  &:hover &_shadow {
+    transform: translateY(4px);
+    transition: transform 250ms cubic-bezier(0.3, 0.7, 0.4, 1.5);
+  }
+
+  &:active &_shadow {
+    transform: translateY(1px);
+    transition: transform 34ms;
+  }
+
+  &:focus-visible {
+    outline: revert;
+    outline: var(--cl-border-outline);
+    outline-offset: 4px;
+  }
+}
+
+.flat {
+  @include typography.font-style(core, base);
+  color: var(--cl-font-color-base);
+  border: none;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: var(--cl-gap-small);
+  padding: var(--cl-gap-medium) var(--cl-gap-big);
+  border-radius: var(--cl-block-border-radius);
+  background: transparent;
+
+  &__extended {
+    width: 100%;
+  }
+
+  &:hover,
+  &:active {
+    background: var(--cl-background-color-light);
+    transition: all 0.3s ease 0s;
+    color: var(--cl-font-color-accent);
+  }
+
+  &:focus-visible {
+    outline: var(--cl-border-outline);
+  }
+
+  &[disabled] {
+    @include mixins.disabled();
+  }
+}
+
+.icon {
+  @include typography.font-style(core, base);
+  color: var(--cl-font-color-base);
+  border: none;
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: var(--cl-gap-medium);
+  border-radius: var(--cl-block-border-radius);
+  background: transparent;
+
+  &:hover {
+    background-color: var(--cl-background-color-light);
+    color: var(--cl-font-color-accent);
+    transition: all 0.3s ease 0s;
+  }
+
+  &:focus-visible {
+    outline: var(--cl-border-outline);
+  }
+
+  &[disabled] {
+    pointer-events: none;
+    opacity: 0.5;
+  }
+
+  svg {
+    height: var(--cl-icon-size);
+    width: var(--cl-icon-size);
+  }
+}
