+++ src/components/modals/alert/Alert.module.scss
@@ -16,12 +16,12 @@
 
 .content {
   @include typography.font-style(core, base);
+  color: var(--cl-font-color-base);
 
   display: flex;
   flex-direction: column;
   gap: var(--cl-gap-medium);
   background: var(--cl-background-color-base);
-  color: var(--cl-font-color-base);
   width: 90vw;
   padding: var(--cl-gap-medium);
   box-shadow: var(--cl-box-shadow);
