+++ src/components/modals/dialog/Dialog.module.scss
@@ -1,0 +1,20 @@
+@use '/src/styles/animations';
+
+.dialog {
+  @include animations.fade-in();
+
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: inherit;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: var(--cl-background-color-transparent);
+
+  &__closing {
+    @include animations.fade-out();
+  }
+}
