+++ src/components/modals/dialog/elements/Header/Header.module.scss
@@ -1,0 +1,13 @@
+@use '/src/styles/typography';
+
+.header {
+  @include typography.font-style(core, header, bold);
+
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: var(--cl-gap-big);
+  padding: var(--cl-gap-big);
+  border-bottom: var(--cl-border-base);
+  color: var(--cl-font-color-base);
+}
